Move bulk blob delete utility from JCloudsArtifactManager to JCloudsVirtualFile.
